버셀 테스트용

+++ src/components/SingleMyPost/SingleMyPost.jsx
@@ -15,6 +15,7 @@ const SingleMyPost = ({
   image,
   comment,
   boardType,
+  setError,
 }) => {
   const [isLoading, setIsLoading] = useState(false);
   const [isDeleted, setIsDeleted] = useState(false); //삭제 누르는 순간 화면에서 바로 삭제를 위해
@@ -22,24 +23,25 @@ const SingleMyPost = ({
   const navigate = useNavigate();
 
   const deletePost = async () => {
-    setIsLoading(true);
-    try {
-      const response = await deleteData(
-        DELETE_MY_POST(boardType, postId),
-        {
-          Authorization: `${localStorage.getItem('grantType')} ${localStorage.getItem('AToken')}`,
-        },
-        {},
-      );
-      console.log(response);
-      if (response.status === 200) {
-        console.log('성공');
-        setIsDeleted(true); // 삭제 상태 업데이트
+    const confirmDelete = window.confirm('해당 글을 삭제하시겠습니까?');
+    if (confirmDelete) {
+      setIsLoading(true);
+      try {
+        const response = await deleteData(
+          DELETE_MY_POST(boardType, postId),
+          {
+            Authorization: `${localStorage.getItem('grantType')} ${localStorage.getItem('AToken')}`,
+          },
+          {},
+        );
+        if (response.status === 200) {
+          setIsDeleted(true); // 삭제 상태 업데이트
+        }
+      } catch (error) {
+        setError(true);
+      } finally {
+        setIsLoading(false);
       }
-    } catch (error) {
-      console.error('delete error:', error);
-    } finally {
-      setIsLoading(false);
     }
   };
 
@@ -63,7 +65,14 @@ const SingleMyPost = ({
       <s.PostContainer>
         <s.Top>
           <s.Title>{title}</s.Title>
-          <s.Time>{formatTime}</s.Time>
+          <s.Delete
+            onClick={(e) => {
+              e.stopPropagation(); // 이벤트 버블링 방지
+              deletePost();
+            }}
+          >
+            삭제
+          </s.Delete>
         </s.Top>
 
         <s.ContentText>{content}</s.ContentText>
@@ -81,14 +90,6 @@ const SingleMyPost = ({
           <span style={{ color: '#92A5BC' }}>{comment}</span>
         </s.Info>
       </s.PostContainer>
-      <s.Delete
-        onClick={(e) => {
-          e.stopPropagation(); // 이벤트 버블링 방지
-          deletePost();
-        }}
-      >
-        삭제
-      </s.Delete>
     </s.PostWrapper>
   );
 };
